+++ src/Entities/InlineKeyboardButton.php
@@ -22,12 +22,14 @@ use Longman\TelegramBot\Exception\TelegramException;
  * @method string getCallbackData()                 Optional. Data to be sent in a callback query to the bot when button is pressed, 1-64 bytes
  * @method string getSwitchInlineQuery()            Optional. If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified inline query in the input field. Can be empty, in which case just the bot’s username will be inserted.
  * @method string getSwitchInlineQueryCurrentChat() Optional. If set, pressing the button will insert the bot‘s username and the specified inline query in the current chat's input field. Can be empty, in which case only the bot’s username will be inserted.
+ * @method string getPay()                          Optional. Specify True, to send a Pay button.
  *
  * @method $this setText(string $text)                                                     Label text on the button
  * @method $this setUrl(string $url)                                                       Optional. HTTP url to be opened when button is pressed
  * @method $this setCallbackData(string $callback_data)                                    Optional. Data to be sent in a callback query to the bot when button is pressed, 1-64 bytes
  * @method $this setSwitchInlineQuery(string $switch_inline_query)                         Optional. If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified inline query in the input field. Can be empty, in which case just the bot’s username will be inserted.
  * @method $this setSwitchInlineQueryCurrentChat(string $switch_inline_query_current_chat) Optional. If set, pressing the button will insert the bot‘s username and the specified inline query in the current chat's input field. Can be empty, in which case only the bot’s username will be inserted.
+ * @method $this setPay(bool $pay)                                                         Optional. Specify True, to send a Pay button.
  */
 class InlineKeyboardButton extends KeyboardButton
 {
@@ -45,7 +47,8 @@ class InlineKeyboardButton extends KeyboardButton
                    array_key_exists('url', $data) ||
                    array_key_exists('callback_data', $data) ||
                    array_key_exists('switch_inline_query', $data) ||
-                   array_key_exists('switch_inline_query_current_chat', $data)
+                   array_key_exists('switch_inline_query_current_chat', $data) ||
+                   array_key_exists('pay', $data)
                );
     }
 
@@ -60,14 +63,14 @@ class InlineKeyboardButton extends KeyboardButton
 
         $num_params = 0;
 
-        foreach (['url', 'callback_data', 'switch_inline_query', 'switch_inline_query_current_chat'] as $param) {
+        foreach (['url', 'callback_data', 'switch_inline_query', 'switch_inline_query_current_chat', 'pay'] as $param) {
             if (!empty($this->getProperty($param))) {
                 $num_params++;
             }
         }
 
         if ($num_params !== 1) {
-            throw new TelegramException('You must use only one of these fields: url, callback_data, switch_inline_query, switch_inline_query_current_chat!');
+            throw new TelegramException('You must use only one of these fields: url, callback_data, switch_inline_query, switch_inline_query_current_chat, pay!');
         }
     }
 
@@ -77,8 +80,8 @@ class InlineKeyboardButton extends KeyboardButton
     public function __call($method, $args)
     {
         // Only 1 of these can be set, so clear the others when setting a new one.
-        if (in_array($method, ['setUrl', 'setCallbackData', 'setSwitchInlineQuery', 'setSwitchInlineQueryCurrentChat'], true)) {
-            unset($this->url, $this->callback_data, $this->switch_inline_query, $this->switch_inline_query_current_chat);
+        if (in_array($method, ['setUrl', 'setCallbackData', 'setSwitchInlineQuery', 'setSwitchInlineQueryCurrentChat', 'setPay'], true)) {
+            unset($this->url, $this->callback_data, $this->switch_inline_query, $this->switch_inline_query_current_chat, $this->pay);
         }
 
         return parent::__call($method, $args);